adding scan to recipes

+++ recipemonkeyapp/views/recipe.py
@@ -61,15 +61,17 @@ def scan(request, id):
 		if formset.is_valid():
 			for form in formset:
 				si=form.save(commit=False)
-				si.content_object=i
-				si.object_id=i.id
-				si.content_type=ContentType.objects.get_for_model(i)
+				
+				if si.id:
+			        si.content_object=i
+        			si.object_id=i.id
+        			si.content_type=ContentType.objects.get_for_model(i)
 
-				if si.quantity > 0:
-					si.save()
-					
-				if si.quantity <=0:
-				    si.delete()
+        			if si.quantity > 0:
+        				si.save()
+				
+        			if si.quantity <=0:
+        			    si.delete()
 
 			return redirect('recipemonkeyapp.views.recipe.scan',id=i.id)
